fix: make fs r/w working dir independent
