+++ type-generator/src/schema/extract.rs
@@ -1,6 +1,7 @@
 use crate::build_errors::BuildError;
+use crate::utils::read_write_fs::read_file_to_string;
 use serde_json::Value;
-use std::fs;
+use std::path::Path;
 
 /// Loads a JSON schema from disk, optionally applies a patch function,
 /// and returns the parsed `RootSchema`.
@@ -8,8 +9,7 @@ pub fn load_and_patch_schema(
     input: &str,
     schema_patch: Option<&dyn Fn(&mut Value)>,
 ) -> Result<schemars::schema::RootSchema, BuildError> {
-    println!("Building types from schema: {input}");
-    let content = fs::read_to_string(input)?;
+    let content = read_file_to_string(Path::new(input))?;
     let mut schema_value: Value = serde_json::from_str(&content)?;
 
     if let Some(patch_fn) = schema_patch {
