+++ type-generator/src/schema/generate.rs
@@ -1,6 +1,6 @@
 use crate::build_errors::BuildError;
 use crate::utils::codegen_snippets::{GENERATED_CODE_HEADER, add_ignore_clippy, add_ignore_rustfmt};
-use crate::utils::write_to_fs::write_generated_file;
+use crate::utils::read_write_fs::write_generated_file;
 use typify::{TypeSpace, TypeSpaceSettings};
 
 pub fn generate_from_schema(
