+++ type-generator/src/utils/mod.rs
@@ -1,2 +1,2 @@
 pub mod codegen_snippets;
-pub mod write_to_fs;
+pub mod read_write_fs;
